리펙토링- "AudioGuideBLEControl.kt부분 함수"connectBluetoothGatt"부분과"sendDataToCharacteristic"부분 AudioGuideUtils.kt 쪽으로 이전 및 설명 추가 "AudioGuideBLEConnectActivity.kt" 중 navigateToNextActivity 부분 AudioGuideUtils.kt쪽으로 이전 및 설명 추가
